requirements.txt
fastapi
uvicorn
websockets
pydantic
requests
langchain
langchain-google-vertexai
google-cloud-aiplatform